Adds user-facing telemetry submission.
By default Pulp systems will now submit anonymous telemetry information
to https://analytics.pulpproject.org/. It documents exactly what is
submitted, and introduces the `TELEMETRY` setting which can be used to
disable anonymous telemetry submission entirely.

closes #3115
