+++ CHANGES/3115.feature
@@ -1,0 +1,3 @@
+Introduces anonymous telemetry data posting to `<https://analytics.pulpproject.org/>`_. This is
+enabled by default, and can be disabled by setting the ``TELEMETRY`` setting to ``False``. See the
+:ref:`telemetry docs <telemetry>` for more info on exactly what is posted along with an example.
